+++ task-4/Code/src/Main.java
@@ -1,0 +1,7 @@
+import autoservice.manager.impl.ServiceManager;
+
+public class Main {
+    public static void main(String[] args) {
+        ServiceManager manager = new ServiceManager();
+    }
+}
