+++ task-4/Code/src/autoservice/models/garage/garageStatus/GarageStatue.java
@@ -1,0 +1,6 @@
+package autoservice.models.garage.garageStatus;
+
+public enum GarageStatue {
+    AVAILABLE,
+    OCCUPIED
+}
