+++ task-4/Code/src/autoservice/models/garagePlace/GaragePlace.java
@@ -1,0 +1,26 @@
+package autoservice.models.garagePlace;
+
+public class GaragePlace {
+    private int placeNumber;
+    private boolean isOccupied = false;
+
+    public GaragePlace(int placeNumber) {
+        this.placeNumber = placeNumber;
+    }
+
+    public int getPlaceNumber() {
+        return this.placeNumber;
+    }
+
+    public void setPlaceNumber(int placeNumber) {
+        this.placeNumber = placeNumber;
+    }
+
+    public boolean isOccupied() {
+        return this.isOccupied;
+    }
+
+    public void setOccupied(boolean isOccupied) {
+        this.isOccupied = isOccupied;
+    }
+}
