+++ task-4/Code/src/autoservice/models/master/masterStatus/MasterStatus.java
@@ -1,0 +1,6 @@
+package autoservice.models.master.masterStatus;
+
+public enum MasterStatus {
+    AVAILABLE,
+    OCCUPIED
+}
