+++ task-4/Code/src/autoservice/models/order/orderStatus/OrderStatus.java
@@ -1,0 +1,8 @@
+package autoservice.models.order.orderStatus;
+
+public enum OrderStatus {
+    CREATED,
+    COMPLETED,
+    CANCELLED,
+    IN_PROGRESS
+}
