+++ task-4/Code/src/autoservice/servicesSorting/DataSort/DataSortInterface.java
@@ -1,0 +1,10 @@
+package autoservice.servicesSorting.DataSort;
+
+import java.time.LocalDateTime;
+
+public interface DataSortInterface {
+
+    int getFreePlacesOnDate(LocalDateTime date);
+
+    LocalDateTime getNearestFreeDate();
+}
