+++ task-4/Code/src/autoservice/servicesSorting/GaragePlacesSort/GaragePlacesSortInterface.java
@@ -1,0 +1,10 @@
+package autoservice.servicesSorting.GaragePlacesSort;
+
+import autoservice.models.garagePlace.GaragePlace;
+
+import java.util.List;
+
+public interface GaragePlacesSortInterface {
+
+    List<GaragePlace> getAvailableGaragePlaces();
+}
