+++ task-4/Code/src/autoservice/servicesSorting/MastersSort/MastersSortInterface.java
@@ -1,0 +1,14 @@
+package autoservice.servicesSorting.MastersSort;
+
+import autoservice.models.master.Master;
+import autoservice.models.order.Order;
+
+import java.util.Comparator;
+import java.util.List;
+
+public interface MastersSortInterface {
+
+    List<Master> getMastersByOrders(Order order);
+
+    List<Master> getSortedMasters(List<Comparator<Master>> comparators);
+}
