--- task-4/Code/src/main/java/autoservice/models/garage/essence/master/masterStatus/MasterStatus.java
@@ -1,6 +1,0 @@
-package main.java.autoservice.models.garage.essence.master.masterStatus;
-
-public enum MasterStatus {
-    AVAILABLE,
-    OCCUPIED
-}
