--- task-4/Code/src/main/java/autoservice/models/garage/garageStatus/GarageStatue.java
@@ -1,6 +1,0 @@
-package main.java.autoservice.models.garage.garageStatus;
-
-public enum GarageStatue {
-    AVAILABLE,
-    OCCUPIED
-}
