--- task-4/Code/src/main/java/autoservice/models/order/orderStatus/OrderStatus.java
@@ -1,8 +1,0 @@
-package main.java.autoservice.models.order.orderStatus;
-
-public enum OrderStatus {
-    CREATED,
-    COMPLETED,
-    CANCELLED,
-    IN_PROGRESS
-}
